requirements.txt
# ============================================
# PRODUCTION REQUIREMENTS - Sistema SaaS Completo
# ============================================

# Base Framework
crewai>=0.28.8
crewai-tools>=0.0.20
flask>=3.0.0
flask-cors>=4.0.0
langchain-openai>=0.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Autenticação e Segurança
pyjwt==2.8.0
flask-jwt-extended==4.5.3
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Rate Limiting e Cache
redis==5.0.0
flask-limiter==3.5.0
flask-caching==2.1.0

# Async e Queue
celery==5.3.4
celery[redis]==5.3.4
flower==2.0.1
kombu==5.3.4

# Pagamentos (comentado - instalar manualmente se disponivel)
# abacatepay==1.0.0

# Monitoring e Logging
prometheus-client==0.19.0
sentry-sdk[flask]==1.39.2
python-json-logger==2.0.7

# Validação e Parsing
pydantic>=2.11.9
marshmallow==3.20.1
email-validator==2.1.0

# HTTP e Requests
requests==2.31.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
sqlalchemy==2.0.25

# Data Processing (Pipeline Leilao)
pandas==2.1.4
numpy==1.26.2

# PDF Generation
reportlab==4.0.8

# Scheduling
schedule==1.2.1
APScheduler==3.10.4

# Retry Logic
tenacity==8.2.3
